+++ CHANGES.md
@@ -1,5 +1,8 @@
 ## dev
 
+* Focus new iframe widget (type ``` to test with code block)
+* NEW -- optional `options.onDropFileOnBlock(id, file)` to a block's cover
+
 ## 0.14.0 - 2016-06-06
 
 * NEW -- optional `options.onDropFiles(index, files)` to drop files into editor flow